Use ApiService for recording and report downloads
ApiService includes credentials on requests and downloads files into
browser memory, then produces a virtual anchor link and "clicks" it,
allowing the request filename to also be modified. This is used for
downloading both JFR binary recordings as well as HTML reports for the
user to store locally (or for reports, to open in a new browser tab if
they wish)
